automation/requirements.txt
python-dotenv
httpx
aiofiles
todoist-api-python

